restructure action ymls
