+++ .github/workflows/sapporo.yml
@@ -1,10 +1,10 @@
-name: Ruby CI
+name: Sapporo CI
 
 on:
   pull_request:
     branches: [main]
   push:
-    branches: [main]
+    branches: [main, dev] #TODO: remove dev from here
 
 jobs:
   test:
@@ -46,9 +46,31 @@ jobs:
           cache: "pnpm"
       - name: Install dependencies
         run: pnpm install
+      - name: Download Sapporo
+        uses: actions/checkout@v3
+        with:
+          repository: phac-nml/sapporo-service
+          ref: irida-next-chown
+          path: ./sapporo
+      - name: Build Sapporo
+        run: |
+          cd sapporo
+          IRIDA_NEXT_PATH=${GITHUB_WORKSPACE}/tmp MYUID="$(id -u)" MYGID="$(id -g)" docker compose -f compose.irida-next.yml up -d --build
+          cd ..
+      - name: Run Sapporo
+        uses: JarvusInnovations/background-action@v1
+        with:
+          run: |
+            IRIDA_NEXT_PATH=${GITHUB_WORKSPACE}/tmp MYUID="$(id -u)" MYGID="$(id -g)" docker compose -f compose.irida-next.yml exec app bash -c "sapporo"
+          wait-on:
+            http://localhost:1122/service-info
+          wait-for: 2m
+          working-directory: sapporo
+      - name: Update permissions
+        run: sudo chmod -R 777 ${GITHUB_WORKSPACE}
       - name: Run tests
         run: |
-          bin/rails test:all --except "/test_integration_sapporo*/"
+          bin/rails test test/integration/sapporo.rb
       - name: Report coverage
         if: github.event_name == 'pull_request'
         uses: aki77/simplecov-report-action@v1
